feat: :sparkles: retrieve gitmojis from the internet
In this commit, we add the ability to retrieve the latest gitmojis from the internet.
We also start to implement the ability to load the gitmoji list from a local file, instead.
The main reason for this is to not download the gitmoji list everytime the CLI tool is started (as this may take some time on slower connections).